Add files via upload
This esp32 code is work in progress. Initially it is in the BLE mode and scans for 4 BLE devices whose id starts with ff:ff: Once it acquires the name, id, it stops the BLE mode and connects to Wifi and writes the information into google sheets. That is all it does for now. 

Issues : If anyone in the bluetooth vicinity has the blutooth app that comes with the BLE tags bought on Amazon, then it is possible that the BLE tage will get connected to the smatphone and they would not be transmitting the name, id etc information. Need to find a solution to this issue, for you cannot enforce that no one has these apps.